requirements/min.txt
# Mini needed in Development
Flask==3.1.2
APIFlask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Cors==5.0.1
boto3>=1.36,<2.0
python-dotenv==1.1.0
ec2-metadata==2.17.0
